- #129: Moved association-modal feature out of instance
