+++ client/edition/association-modal/browse-selected-entities.js
