+++ client/edition/association-modal/initialize-selected-entities.js
