+++ client/edition/association-modal/selected-details.hbs
