+++ client/edition/association-modal/update-selected-details.js
