+++ client/edition/instance/post-render.js
@@ -5,7 +5,7 @@ const actionLink = require('./action-link');
 const actionGoto = require('./../../shared/action-goto');
 const addChild = require('./add-child');
 const addSibling = require('./add-sibling');
-const associationModal = require('./association-modal');
+const associationModal = require('./../association-modal');
 const carouselPagination = require('./carousel-pagination');
 const ChangeLogs = require('./../change-logs');
 const deleteEntity = require('./delete-entity');
